% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_utils.R
\name{calculatePooledEstimate}
\alias{calculatePooledEstimate}
\title{Calculates the Inverse Variance Pooled Estimate Including Null Folds}
\usage{
calculatePooledEstimate(results_df, n_folds, delta = NULL)
}
\arguments{
\item{results_df}{Table of results}

\item{n_folds}{Total number of folds specified}
}
\description{
Does a weighted combination estimate for folds with estimates
and null folds finding 0 using inverse variance
}
